Break up e2e tests
