+++ frontend/e2e/xmas-tree-info.po.ts
@@ -1,0 +1,86 @@
+import { browser, element, by } from 'protractor';
+
+export class TreesSidebarPage {
+  navigateTo(forestId) {
+    return browser.get('/christmas-trees/forests/' + forestId + '/tree-guidelines');
+  }
+
+  getTreeSelectionLink() {
+    return element(by.id('tree-selection-link'));
+  }
+
+  getTreeLocationLink() {
+    return element(by.id('tree-locations-link'));
+  }
+
+  getTreeSpecies(type, index) {
+    return element(by.id(`tree-${type}-species-${index}`));
+  }
+
+  cuttingAreaDates(index) {
+    return element(by.id(`tree-cutting-areas-dates-${index}`));
+  }
+
+  cuttingAreaHours(index) {
+    return element(by.id(`tree-cutting-areas-hours-${index}`));
+  }
+
+  contactUsSectionLink() {
+    return element(by.id('contact-information-link'));
+  }
+
+  contactUsSection() {
+    return element(by.id('contact-information'));
+  }
+
+  contactUsHeadquarters() {
+    return element(by.id('tree-contact-headquarters-0'));
+  }
+
+  contactUsDistrict(id) {
+    return element(by.id(`tree-contact-district-${id}`));
+  }
+
+  cuttingDatesSectionLink() {
+    return element(by.id('cutting-dates-link'));
+  }
+
+  cuttingDatesSection() {
+    return element(by.id('cutting-dates'));
+  }
+
+  cuttingDatesSeasonStartAndEnd() {
+    return element(by.id('cutting-season-start-end-date'));
+  }
+
+  cuttingDatesHoursDefault() {
+    return element(by.id('cutting-season-daily-hours-default'));
+  }
+
+  whenToCutDatesSection() {
+    return element(by.id('cutting-dates'));
+  }
+
+  whenToCutDatesSectionLink() {
+    return element(by.id('cutting-dates-link'));
+  }
+
+  rulesToKnowSectionLink() {
+    return element(by.id('need-to-know-link'));
+  }
+
+  buyPermitLink() {
+    return element(by.id('buy-permit-link'));
+  }
+
+  rulesToKnowSection() {
+    return element(by.id('need-to-know'));
+  }
+}
+
+export class TreesForestFinderPage {
+  navigateTo() {
+    return browser.get('/christmas-trees/forests');
+  }
+}
+
