+++ frontend/e2e/xmas-tree-locations.e2e-spec.ts
@@ -1,13 +1,16 @@
-import { TreesSidebarPage } from './app.po';
+import { TreesSidebarPage } from './xmas-tree-info.po';
 import { browser, element, by, Key, protractor } from 'protractor';
 
 describe('Xmas tree  - Where to Find Your Tree page', () => {
   let page: TreesSidebarPage;
 
-  describe( 'Mt Hood', () => {
+  describe('Mt Hood', () => {
     beforeEach(() => {
       page = new TreesSidebarPage();
-      browser.driver.manage().window().setSize(1400, 900);
+      browser.driver
+        .manage()
+        .window()
+        .setSize(1400, 900);
       page.navigateTo('mthood');
       browser.sleep(800);
       page.getTreeLocationLink().click();
@@ -15,19 +18,17 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
     });
 
     it('should have a link to the external forest maps-pubs', () => {
-      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual('https://www.fs.usda.gov/main/mthood/maps-pubs');
+      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual(
+        'https://www.fs.usda.gov/main/mthood/maps-pubs'
+      );
     });
 
     it('should have a where to find a tree section link', () => {
-      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual(
-        'Where to find your tree'
-      );
+      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual('Where to find your tree');
     });
 
     it('should have a cutting areas sub-section link', () => {
-      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual(
-        'Cutting area maps'
-      );
+      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual('Cutting area maps');
     });
 
     it('should have a prohibited areas sub-section link', () => {
@@ -40,13 +41,15 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
       expect<any>(element(by.id('tree-locations-maps-0')).isPresent()).toBeTruthy();
       expect<any>(element(by.id('tree-locations-maps-1')).isPresent()).toBeTruthy();
     });
-
   });
 
-  describe( 'Shoshone', () => {
+  describe('Shoshone', () => {
     beforeEach(() => {
       page = new TreesSidebarPage();
-      browser.driver.manage().window().setSize(1400, 900);
+      browser.driver
+        .manage()
+        .window()
+        .setSize(1400, 900);
       page.navigateTo('shoshone');
       browser.sleep(800);
       page.getTreeLocationLink().click();
@@ -54,20 +57,17 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
     });
 
     it('should have a link to the external forest maps-pubs', () => {
-      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual('https://www.fs.usda.gov/main/shoshone/maps-pubs');
+      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual(
+        'https://www.fs.usda.gov/main/shoshone/maps-pubs'
+      );
     });
 
-
     it('should have a where to find a tree section link', () => {
-      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual(
-        'Where to find your tree'
-      );
+      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual('Where to find your tree');
     });
 
     it('should have a cutting areas sub-section link', () => {
-      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual(
-        'Cutting area maps'
-      );
+      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual('Cutting area maps');
     });
 
     it('should have a prohibited areas sub-section link', () => {
@@ -75,13 +75,15 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
         'Prohibited areas and restrictions'
       );
     });
-
   });
 
-  describe( 'Arapaho/Roosevelt', () => {
+  describe('Arapaho/Roosevelt', () => {
     beforeEach(() => {
       page = new TreesSidebarPage();
-      browser.driver.manage().window().setSize(1400, 900);
+      browser.driver
+        .manage()
+        .window()
+        .setSize(1400, 900);
       page.navigateTo('arp');
       browser.sleep(800);
       page.getTreeLocationLink().click();
@@ -89,20 +91,17 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
     });
 
     it('should have a link to the external forest maps-pubs', () => {
-      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual('https://www.fs.usda.gov/main/arp/maps-pubs');
+      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual(
+        'https://www.fs.usda.gov/main/arp/maps-pubs'
+      );
     });
 
-
     it('should have a where to find a tree section link', () => {
-      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual(
-        'Where to find your tree'
-      );
+      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual('Where to find your tree');
     });
 
     it('should have a cutting areas sub-section link', () => {
-      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual(
-        'Cutting area maps'
-      );
+      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual('Cutting area maps');
     });
 
     it('should have a prohibited areas sub-section link', () => {
@@ -117,10 +116,13 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
     });
   });
 
-  describe( 'Flathead', () => {
+  describe('Flathead', () => {
     beforeEach(() => {
       page = new TreesSidebarPage();
-      browser.driver.manage().window().setSize(1400, 900);
+      browser.driver
+        .manage()
+        .window()
+        .setSize(1400, 900);
       page.navigateTo('flathead');
       browser.sleep(900);
       page.getTreeLocationLink().click();
@@ -128,19 +130,17 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
     });
 
     it('should have a link to the external forest maps-pubs', () => {
-      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual('https://www.fs.usda.gov/main/flathead/maps-pubs');
+      expect<any>(element(by.id('forest-maps-publications-link')).getAttribute('href')).toEqual(
+        'https://www.fs.usda.gov/main/flathead/maps-pubs'
+      );
     });
 
     it('should have a where to find a tree section link', () => {
-      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual(
-        'Where to find your tree'
-      );
+      expect<any>(element(by.id('tree-locations-link')).getText()).toEqual('Where to find your tree');
     });
 
     it('should have a cutting areas sub-section link', () => {
-      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual(
-        'Cutting area maps'
-      );
+      expect<any>(element(by.id('tree-locations-allowed-link')).getText()).toEqual('Cutting area maps');
     });
 
     it('should have a prohibited areas sub-section link', () => {
@@ -148,8 +148,5 @@ describe('Xmas tree  - Where to Find Your Tree page', () => {
         'Prohibited areas and restrictions'
       );
     });
-
   });
-
-
 });
